backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
python-multipart
pytesseract
pillow
reportlab
pydantic
